import dojo-examples from sandbox
The dojo-examples code had been moved to the attic.

Since we've added dojo back into core-1.4.x we also need the examples.

This is a --squash commit of the filtered-branch from sandbox SHA1:
e7267068a99daf4556e00269fc270af4a6850f65

which is one commit before it was moved into the attic.