more dojo changes
Creates a dojo-examples module that is a copy of the dojo-original-examples module but with the
dependency switched to dojo-api.

Get rid of the dojo-resources module as dojotoolkit.org distributes the zip through maven.

See the dojo-api/pom.xml for how it is automatically downloaded into the target/classes/dojo directory.

Change the implementation-version for the dojo-api.jar to hold the dojo.version instead of the wicket version so that within Java
the proper path to dojo can be constructed.

Note the dojo-examples doesn't work as it needs to be adapted to the more modern API.